pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tongyi-cli-interactive"
version = "0.1.0"
description = "Interactive CLI for Tongyi Agent - Modern terminal interface with rich UI, session management, and tool integration"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Tongyi CLI Contributors"}
]
keywords = ["cli", "interactive", "tongyi", "agent", "research", "assistant", "terminal", "ui"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Shells",
    "Topic :: Terminals",
]
dependencies = [
    "pydantic>=2.0",
    "requests>=2.28",
    "pandas>=2.0",
    "markdown>=3.5",
    "PyYAML>=6.0",
    "rich>=13.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0",
    "click>=8.0.0",
    "tabulate>=0.9.0",
    "colorama>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
]
claude-sdk = [
    "claude-agent-sdk>=0.1.0",
    "nest-asyncio>=1.5.0",
]
agent-lightning = [
    "agentlightning>=0.1.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
]

[project.urls]
Homepage = "https://github.com/your-org/tongyi-cli-interactive"
Repository = "https://github.com/your-org/tongyi-cli-interactive"
Issues = "https://github.com/your-org/tongyi-cli-interactive/issues"
Documentation = "https://github.com/your-org/tongyi-cli-interactive/blob/main/CLI_GUIDE.md"

[project.scripts]
tongyi = "tongyi_agent.cli:main"
tongyi-cli = "tongyi_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["tongyi_agent*", "src*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]
